:sparkles: Feat: 즐겨찾기 생성 및 조회 data class 정의

+++ app/src/main/java/com/example/snack4diet/api/addBookmark/AddBookmark.kt
@@ -2,7 +2,7 @@ package com.example.snack4diet.api.addBookmark
 
 data class AddBookmark(
     val baseNutrition: BaseNutrition,
-    val foodId: Int,
+    val foodId: Long,
     val name: String,
-    val userId: Int
+    val userId: Long
 )
